// types/generated/sanity.types.ts
export type TestimonialFeatured = {
  _type: 'testimonial.featured';
  testimonial?: {
    _ref: string;
    _type: 'reference';
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: 'testimonial';
  };
};
// ...
export type PricingList = {
  _type: 'pricing-list';
  intro?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  tiers?: Array<{
    _ref: string;
    _type: 'reference';
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: 'pricing';
  }>;
};

export type LogoList = {
  _type: 'logo-list';
  pretitle?: string;
  intro?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  logos?: Array<{
    _ref: string;
    _type: 'reference';
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: 'logo';
  }>;
  logoType?: 'default' | 'light' | 'dark';
  autoScroll?: boolean;
};

export type HeroSplit = {
  _type: 'hero.split';
  pretitle?: string;
  content?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  ctas?: Array<
    {
      _key: string;
    } & Cta
  >;
  image?: {
    asset?: {
      _ref: string;
      _type: 'reference';
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    onRight?: boolean;
    loading?: 'lazy' | 'eager';
    _type: 'image';
  };
};

export type HeroSaas = {
  _type: 'hero.saas';
  pretitle?: string;
  content?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  ctas?: Array<
    {
      _key: string;
    } & Cta
  >;
  image?: {
    asset?: {
      _ref: string;
      _type: 'reference';
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    faded?: boolean;
    loading?: 'lazy' | 'eager';
    _type: 'image';
  };
};

export type Hero = {
  _type: 'hero';
  pretitle?: string;
  content?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  ctas?: Array<
    {
      _key: string;
    } & Cta
  >;
  bgType?: 'image' | 'video';
  bgImage?: {
    asset?: {
      _ref: string;
      _type: 'reference';
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    loading?: 'lazy' | 'eager';
    overlay?: boolean;
    _type: 'image';
  };
  bgImageMobile?: {
    asset?: {
      _ref: string;
      _type: 'reference';
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: 'image';
  };
  bgVideo?: {
    asset?: {
      _ref: string;
      _type: 'reference';
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: 'sanity.fileAsset';
    };
    alt?: string;
    overlay?: boolean;
    _type: 'file';
  };
  bgVideoThumbnail?: {
    asset?: {
      _ref: string;
      _type: 'reference';
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: 'image';
  };
  textAlign?: 'left' | 'center' | 'right';
  alignItems?: 'start' | 'center' | 'end';
};

export type FlagList = {
  _type: 'flag-list';
  intro?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  items?: Array<{
    icon?: {
      asset?: {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
      };
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: 'image';
    };
    content?: Array<{
      children?: Array<{
        marks?: Array<string>;
        text?: string;
        _type: 'span';
        _key: string;
      }>;
      style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
      listItem?: 'bullet' | 'number';
      markDefs?: Array<{
        href?: string;
        _type: 'link';
        _key: string;
      }>;
      level?: number;
      _type: 'block';
      _key: string;
    }>;
    _key: string;
  }>;
  iconSize?: number;
  iconPosition?: 'top' | 'left';
};

export type CustomHtml = {
  _type: 'custom-html';
  html?: Code;
};

export type CreativeModule = {
  _type: 'creative-module';
  intro?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  modules?: Array<{
    subModules?: Array<
      | {
          ctas?: Array<
            {
              _key: string;
            } & Cta
          >;
          _type: 'ctas';
          _key: string;
        }
      | {
          icon?: {
            asset?: {
              _ref: string;
              _type: 'reference';
              _weak?: boolean;
              [internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
            };
            hotspot?: SanityImageHotspot;
            crop?: SanityImageCrop;
            _type: 'image';
          };
          size?: number;
          _type: 'icon';
          _key: string;
        }
      | {
          asset?: {
            _ref: string;
            _type: 'reference';
            _weak?: boolean;
            [internalGroqTypeReferenceTo]?: 'sanity.imageAsset';
          };
          hotspot?: SanityImageHotspot;
          crop?: SanityImageCrop;
          alt?: string;
          aspectRatio?: string;
          _type: 'image';
          _key: string;
        }
      | {
          content?: Array<{
            children?: Array<{
              marks?: Array<string>;
              text?: string;
              _type: 'span';
              _key: string;
            }>;
            style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
            listItem?: 'bullet' | 'number';
            markDefs?: Array<{
              href?: string;
              _type: 'link';
              _key: string;
            }>;
            level?: number;
            _type: 'block';
            _key: string;
          }>;
          _type: 'richtext';
          _key: string;
        }
      | ({
          _key: string;
        } & CustomHtml)
    >;
    colSpan?: number;
    _key: string;
  }>;
  columns?: number;
  bordered?: boolean;
  textAlign?: 'left' | 'center' | 'right';
  alignItems?: 'start' | 'center' | 'end';
};

export type Callout = {
  _type: 'callout';
  content?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  ctas?: Array<
    {
      _key: string;
    } & Cta
  >;
};

export type Breadcrumbs = {
  _type: 'breadcrumbs';
  crumbs?: Array<
    {
      _key: string;
    } & Link
  >;
};

export type BlogList = {
  _type: 'blog-list';
  intro?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  layout?: 'grid' | 'carousel';
  limit?: number;
  displayFilters?: boolean;
  predefinedFilters?: Array<{
    _ref: string;
    _type: 'reference';
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: 'blog.category';
  }>;
};

export type AccordionList = {
  _type: 'accordion-list';
  intro?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  items?: Array<{
    summary?: string;
    content?: Array<{
      children?: Array<{
        marks?: Array<string>;
        text?: string;
        _type: 'span';
        _key: string;
      }>;
      style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
      listItem?: 'bullet' | 'number';
      markDefs?: Array<{
        href?: string;
        _type: 'link';
        _key: string;
      }>;
      level?: number;
      _type: 'block';
      _key: string;
    }>;
    open?: boolean;
    _key: string;
  }>;
  layout?: 'vertical' | 'horizontal';
  uid?: string;
};

export type Uid = string;

export type Modules = Array<
  | ({
      _key: string;
    } & AccordionList)
  | ({
      _key: string;
    } & BlogList)
  | ({
      _key: string;
    } & Breadcrumbs)
  | ({
      _key: string;
    } & Callout)
  | ({
      _key: string;
    } & CreativeModule)
  | ({
      _key: string;
    } & CustomHtml)
  | ({
      _key: string;
    } & FlagList)
  | ({
      _key: string;
    } & Hero)
  | ({
      _key: string;
    } & HeroSaas)
  | ({
      _key: string;
    } & HeroSplit)
  | ({
      _key: string;
    } & LogoList)
  | ({
      _key: string;
    } & PricingList)
  | ({
      _key: string;
    } & RichtextModule)
  | ({
      _key: string;
    } & StatList)
  | ({
      _key: string;
    } & StepList)
  | ({
      _key: string;
    } & TestimonialList)
  | ({
      _key: string;
    } & TestimonialFeatured)
>;
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;